add nonlocal interaction to Spookynet layers, test passed
