+++ enerzyme/models/spookynet/interaction.py
@@ -1,7 +1,9 @@
+from typing import Optional
 import torch
 from torch.nn import Module, Linear, init
 from ..functional import segment_sum
 from ..layers.electron_embedding import ResidualMLP
+from ..layers.attention import Attention
 
 
 class LocalInteraction(Module):
@@ -117,7 +119,73 @@ class LocalInteraction(Module):
 
 
 class NonlocalInteraction(Module):
-    ...
+    """
+    Block for updating atomic features through nonlocal interactions with all
+    atoms.
+
+    Arguments:
+        num_features (int):
+            Dimensions of feature space.
+        num_basis_functions (int):
+            Number of radial basis functions.
+        num_residual_pre_i (int):
+            Number of residual blocks applied to atomic features in i branch
+            (central atoms) before computing the interaction.
+        num_residual_pre_j (int):
+            Number of residual blocks applied to atomic features in j branch
+            (neighbouring atoms) before computing the interaction.
+        num_residual_post (int):
+            Number of residual blocks applied to interaction features.
+        activation (str):
+            Kind of activation function. Possible values:
+            'swish': Swish activation function.
+            'ssp': Shifted softplus activation function.
+    """
+
+    def __init__(
+        self,
+        dim_embedding: int,
+        num_residual_q: int,
+        num_residual_k: int,
+        num_residual_v: int,
+        activation: str = "swish",
+    ) -> None:
+        """ Initializes the NonlocalInteraction class. """
+        super(NonlocalInteraction, self).__init__()
+        self.resblock_q = ResidualMLP(
+            dim_embedding, num_residual_q, activation, zero_init=True
+        )
+        self.resblock_k = ResidualMLP(
+            dim_embedding, num_residual_k, activation, zero_init=True
+        )
+        self.resblock_v = ResidualMLP(
+            dim_embedding, num_residual_v, activation, zero_init=True
+        )
+        self.attention = Attention(dim_embedding, dim_embedding)
+        self.reset_parameters()
+
+    def reset_parameters(self) -> None:
+        """ For compatibility with other modules. """
+        pass
+
+    def forward(
+        self,
+        x: torch.Tensor,
+        num_batch: int,
+        batch_seg: torch.Tensor,
+        mask: Optional[torch.Tensor] = None,
+    ) -> torch.Tensor:
+        """
+        Evaluate interaction block.
+        N: Number of atoms.
+
+        x (FloatTensor [N, num_features]):
+            Atomic feature vectors.
+        """
+        q = self.resblock_q(x)  # queries
+        k = self.resblock_k(x)  # keys
+        v = self.resblock_v(x)  # values
+        return self.attention(q, k, v, num_batch, batch_seg, mask)
 
 
 class InteractionModule(Module):
